refactor(project): redesign of the packages.
