+++ src/main/java/br/com/fiapbook/shared/model/entity/BaseEntity.java
@@ -1,0 +1,50 @@
+package br.com.fiapbook.shared.model.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.Version;
+import java.io.Serializable;
+import java.time.OffsetDateTime;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+@MappedSuperclass
+@Data
+@SuperBuilder
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(onlyExplicitlyIncluded = true)
+public abstract class BaseEntity implements Serializable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    @EqualsAndHashCode.Include
+    protected UUID id;
+
+    @Builder.Default
+    @Column(nullable = false)
+    protected Boolean deleted = false;
+
+    @Version
+    @Column(nullable = false)
+    protected Long version;
+
+    @CreationTimestamp
+    @Column(nullable = false)
+    protected OffsetDateTime created_at;
+
+    @UpdateTimestamp
+    @Column(nullable = false)
+    protected OffsetDateTime updated_at;
+
+}
