+++ src/main/java/br/com/fiapbook/user/application/usecase/GetAllUsersUseCase.java
@@ -1,0 +1,21 @@
+package br.com.fiapbook.user.application.usecase;
+
+import br.com.fiapbook.user.model.entity.User;
+import br.com.fiapbook.user.model.service.UserService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Component;
+
+@Component
+public class GetAllUsersUseCase {
+
+  private final UserService userService;
+
+  public GetAllUsersUseCase(UserService userService) {
+    this.userService = userService;
+  }
+
+  public Page<User> execute(Pageable pageable){
+    return userService.getAllUsersPaginated(pageable);
+  }
+}
