+++ src/main/java/br/com/fiapbook/user/model/entity/User.java
@@ -1,0 +1,31 @@
+package br.com.fiapbook.user.model.entity;
+
+import br.com.fiapbook.shared.model.entity.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+@Data
+@EqualsAndHashCode(callSuper = true)
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(schema = "user_management", name = "users")
+public class User extends BaseEntity {
+
+  @Column(nullable = false)
+  private String name;
+
+  @Column(nullable = false)
+  private String email;
+
+  @Column(nullable = false)
+  private String password;
+
+}
