+++ src/main/java/br/com/fiapbook/user/model/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package br.com.fiapbook.user.model.repository;
+
+import br.com.fiapbook.user.model.entity.User;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+
+}
