+++ src/main/java/br/com/fiapbook/user/model/service/UserService.java
@@ -1,0 +1,25 @@
+package br.com.fiapbook.user.model.service;
+
+import br.com.fiapbook.user.model.repository.UserRepository;
+import br.com.fiapbook.user.model.entity.User;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+
+  private final UserRepository userRepository;
+
+  public UserService(UserRepository userRepository) {
+    this.userRepository = userRepository;
+  }
+
+  public User save(User user){
+    return userRepository.save(user);
+  }
+
+  public Page<User> getAllUsersPaginated(Pageable pageable){
+    return userRepository.findAll(pageable);
+  }
+}
